+++ Editor/AvatarOptimizerSettings.cs
@@ -9,51 +9,9 @@ using d4rkpl4y3r.AvatarOptimizer.Extensions;
 
 public class AvatarOptimizerSettings : EditorWindow
 {
-    #pragma warning disable 0414
-    private static bool OptimizeOnUpload = true;
-    private static bool WritePropertiesAsStaticValues = true;
-    private static bool MergeSkinnedMeshes = true;
-    private static bool MergeSkinnedMeshesWithShaderToggle = true;
-    private static bool MergeStaticMeshesAsSkinned = true;
-    private static int ForceMergeBlendShapeMissMatch = 2;
-    private static bool MergeDifferentPropertyMaterials = true;
-    private static bool MergeSameDimensionTextures = true;
-    private static bool MergeBackFaceCullingWithCullingOff = false;
-    private static bool MergeDifferentRenderQueue = false;
-    private static bool OptimizeFXLayer = true;
-    private static bool CombineApproximateMotionTimeAnimations = false;
-    private static bool DisablePhysBonesWhenUnused = true;
-    private static bool MergeSameRatioBlendShapes = true;
-    private static bool KeepMMDBlendShapes = false;
-    private static bool DeleteUnusedComponents = true;
-    private static int DeleteUnusedGameObjects = 2;
-    private static bool UseRingFingerAsFootCollider = false;
-    private static bool ProfileTimeUsed = false;
-    #pragma warning restore 0414
-
+    private static d4rkAvatarOptimizer.Settings defaultSettings = new d4rkAvatarOptimizer.Settings();
+    
     private static readonly string PrefsPrefix = "d4rkpl4y3r_AvatarOptimizer_";
-
-    private static Dictionary<string, string> FieldDisplayName = new Dictionary<string, string>() {
-        {"OptimizeOnUpload", "Optimize on Upload"},
-        {"WritePropertiesAsStaticValues", "Write Properties as Static Values"},
-        {"MergeSkinnedMeshes", "Merge Skinned Meshes"},
-        {"MergeSkinnedMeshesWithShaderToggle", "Use Shader Toggles"},
-        {"MergeStaticMeshesAsSkinned", "Merge Static Meshes as Skinned"},
-        {"ForceMergeBlendShapeMissMatch", "Merge Regardless of Blend Shapes"},
-        {"MergeDifferentPropertyMaterials", "Merge Different Property Materials"},
-        {"MergeSameDimensionTextures", "Merge Same Dimension Textures"},
-        {"MergeBackFaceCullingWithCullingOff", "Merge Cull Back with Cull Off"},
-        {"MergeDifferentRenderQueue", "Merge Different Render Queue"},
-        {"KeepMMDBlendShapes", "Keep MMD Blend Shapes"},
-        {"DeleteUnusedComponents", "Delete Unused Components"},
-        {"DeleteUnusedGameObjects", "Delete Unused GameObjects"},
-        {"OptimizeFXLayer", "Optimize FX Layer"},
-        {"CombineApproximateMotionTimeAnimations", "Combine Motion Time Approximation"},
-        {"DisablePhysBonesWhenUnused", "Disable Phys Bones When Unused"},
-        {"MergeSameRatioBlendShapes", "Merge Same Ratio Blend Shapes"},
-        {"UseRingFingerAsFootCollider", "Use Ring Finger as Foot Collider"},
-        {"ProfileTimeUsed", "Profile Time Used"}
-    };
     
     [MenuItem("Tools/d4rkpl4y3r/Avatar Optimizer Settings")]
     static void Init()
@@ -69,7 +27,7 @@ public class AvatarOptimizerSettings : EditorWindow
         EditorGUILayout.HelpBox("These settings are the default values when adding the Avatar Optimizer component to a model."
             +"You can change them here to suit your needs."
             +"\nA * indicates settings that are different from their default value.", MessageType.Info);
-        var fields = typeof(AvatarOptimizerSettings).GetFields(System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic)
+        var fields = typeof(d4rkAvatarOptimizer.Settings).GetFields(System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public)
             .Where(f => f.FieldType == typeof(bool) || f.FieldType == typeof(int)).ToArray();
         if (GUILayout.Button("Reset to Default"))
         {
@@ -87,19 +45,15 @@ public class AvatarOptimizerSettings : EditorWindow
                 GUILayoutUtility.GetRect(35, 15, GUILayout.Width(35));
                 var value = EditorGUILayout.Toggle(0 != GetValue(field.Name), GUILayout.Width(15)) ? 1 : 0;
                 SetValue(field.Name, value);
-                isDefault = field.GetValue(null).Equals(value != 0);
+                isDefault = field.GetValue(defaultSettings).Equals(value != 0);
             }
             else
             {
                 var value = EditorGUILayout.Popup(GetValue(field.Name), new string[] { "Off", "On", "Auto" }, GUILayout.Width(50));
                 SetValue(field.Name, value);
-                isDefault = field.GetValue(null).Equals(value);
+                isDefault = field.GetValue(defaultSettings).Equals(value);
             }
-            if (!FieldDisplayName.TryGetValue(field.Name, out string displayName))
-            {
-                displayName = field.Name;
-            }
-            EditorGUILayout.LabelField(displayName + (isDefault ? "" : " *"));
+            EditorGUILayout.LabelField(d4rkAvatarOptimizer.GetDisplayName(field.Name) + (isDefault ? "" : " *"));
             EditorGUILayout.EndHorizontal();
         }
         EditorGUILayout.Space();
@@ -108,31 +62,31 @@ public class AvatarOptimizerSettings : EditorWindow
 
     public static int GetValue(string key)
     {
-        var field = typeof(AvatarOptimizerSettings).GetField(key, System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic);
+        var field = typeof(d4rkAvatarOptimizer.Settings).GetField(key, System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public);
         if (field == null)
         {
-            throw new System.ArgumentException("Field " + key + " does not exist in AvatarOptimizerSettings");
+            throw new System.ArgumentException("Field " + key + " does not exist in d4rkAvatarOptimizer.Settings");
         }
         if (!EditorPrefs.HasKey(PrefsPrefix + key))
         {
-            EditorPrefs.SetInt(PrefsPrefix + key, field.FieldType == typeof(bool) ? (bool)field.GetValue(null) ? 1 : 0 : (int)field.GetValue(null));
+            EditorPrefs.SetInt(PrefsPrefix + key, field.FieldType == typeof(bool) ? (bool)field.GetValue(defaultSettings) ? 1 : 0 : (int)field.GetValue(defaultSettings));
         }
         return EditorPrefs.GetInt(PrefsPrefix + key);
     }
 
     public static void SetValue(string key, int value)
     {
-        var field = typeof(AvatarOptimizerSettings).GetField(key, System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic);
+        var field = typeof(d4rkAvatarOptimizer.Settings).GetField(key, System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public);
         if (field == null)
         {
-            throw new System.ArgumentException("Field " + key + " does not exist in AvatarOptimizerSettings");
+            throw new System.ArgumentException("Field " + key + " does not exist in d4rkAvatarOptimizer.Settings");
         }
         EditorPrefs.SetInt(PrefsPrefix + key, value);
     }
 
     public static void ApplyDefaults(d4rkAvatarOptimizer optimizer)
     {
-        var fields = typeof(AvatarOptimizerSettings).GetFields(System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic)
+        var fields = typeof(d4rkAvatarOptimizer.Settings).GetFields(System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public)
             .Where(f => f.FieldType == typeof(bool) || f.FieldType == typeof(int)).ToArray();
         foreach (var field in fields)
         {
@@ -141,7 +95,7 @@ public class AvatarOptimizerSettings : EditorWindow
             {
                 throw new System.ArgumentException("Field " + field.Name + " does not exist in d4rkAvatarOptimizer");
             }
-            targetField.SetValue(optimizer, GetValue(field.Name) != 0);
+            targetField.SetValue(optimizer.settings, GetValue(field.Name));
         }
     }